Update some of the FAQs.
I've updated the "What are the Aplite, Basalt and Chalk platforms?" FAQ to mention Diorite, Emery and Flint as well as making it link to the Hardware Information page where full information is possible.

I've removed the "App Publishing" section entirely, as the only question in it was about publishing PebbleKit iOS apps, and that's not something anyone can or should be doing anymore.

Signed-off-by: Radostin Rusev <mitiku04@gmail.com>