Add node as a dependency for macOS
node/npm is a dependency, and it is listed as such for Ubuntu, but not
for macOS.

I could not get the pebble SDK installed until I had node/npm. Since we
are being directed to use homebrew, and it can install node/npm, do that.

Homebrew seems to install both when using either 'node' or 'npm',
so both are not listed separately.

Signed-off-by: Lincoln Ramsay <a1291762@gmail.com>